fix mobile button display
initially, display was based on %; e.g. 50% down the image. this SHOULD
have been exclusive to the image and not the window; this is the case in
desktop, but for whatever asinine reason, isn't on mobile. This update
essentially hand-calculates the % position in pixels, relative
specifically to the image and not the page, and displays the buttons
according to that.